mv shared code into collections v2 trait
